Implemented the Drag & Drop functionality
If you drag and drop nodes it'll act like it's adding a copy of the old
node. If you successfully copy the node then the old node is removed.

TODO: Whenever we change the form, check if there are response
containers without responses and if so we should delete them.